+++ collections/drop_while_test.ts
@@ -24,7 +24,7 @@ Deno.test("dropWhile() handles negatives", () => {
   assertEquals(actual, []);
 });
 
-Deno.test("dropWhile() handles no mutation", () => {
+Deno.test("dropWhile() doesn't mutate input array", () => {
   const arr = [1, 2, 3, 4, 5, 6];
 
   const actual = dropWhile(arr, (i) => i !== 4);
@@ -40,18 +40,70 @@ Deno.test("dropWhile() handles empty input array returns empty array", () => {
   assertEquals(actual, []);
 });
 
-Deno.test("dropWhile() returns empty array when the last element doesn't match the predicate", () => {
-  const arr = [1, 2, 3, 4];
-
-  const actual = dropWhile(arr, (i) => i !== 4);
-
-  assertEquals(actual, [4]);
-});
-
-Deno.test("dropWhile() returns the same array when all elements match the predicate", () => {
+Deno.test("dropWhile() returns the empty array when all elements match the predicate", () => {
   const arr = [1, 2, 3, 4];
 
   const actual = dropWhile(arr, (i) => i !== 400);
 
   assertEquals(actual, []);
+});
+
+Deno.test("dropWhile() returns full elements when the first element satisfies the predicate", () => {
+  const arr = [1, 2, 3, 4, 5, 6];
+  const actual = dropWhile(arr, (i) => i !== 1);
+
+  assertEquals(actual, [1, 2, 3, 4, 5, 6]);
+});
+
+Deno.test("dropWhile() with (i) => i + 2 !== 6", () => {
+  const arr = [1, 2, 3, 4, 5, 6];
+  const actual = dropWhile(arr, (i) => i + 2 !== 6);
+
+  assertEquals(actual, [4, 5, 6]);
+});
+
+Deno.test("dropWhile() returns empty array when the input is empty", () => {
+  const arr: number[] = [];
+
+  const actual = dropWhile(arr, (i) => i > 4);
+
+  assertEquals(actual, []);
+});
+
+Deno.test("dropWhile() handles a generator", () => {
+  function* gen() {
+    yield 1;
+    yield 2;
+    yield 3;
+    yield 4;
+    yield 5;
+    yield 6;
+  }
+
+  const actual = dropWhile(gen(), (i) => i !== 4);
+  assertEquals(actual, [4, 5, 6]);
+});
+
+Deno.test("(unstable) dropWhile() handles a Set", () => {
+  const set = new Set([1, 2, 3, 4, 5, 6]);
+  const actual = dropWhile(set, (i) => i !== 4);
+  assertEquals(actual, [4, 5, 6]);
+});
+
+Deno.test("(unstable) dropWhile() handles a Map", () => {
+  const map = new Map([
+    ["a", 1],
+    ["b", 2],
+    ["c", 3],
+    ["d", 4],
+    ["e", 5],
+    ["f", 6],
+  ]);
+
+  const actual = dropWhile(map, ([_k, v]) => v !== 4);
+  assertEquals(actual, [
+    ["d", 4],
+    ["e", 5],
+    ["f", 6],
+  ]);
 });
